feat: use cancel process instance endpoint in Operate UI
