+++ operate/client/src/modules/components/Operations/v2/Cancel.tsx
@@ -1,0 +1,110 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+import React, {useState} from 'react';
+import {Modal} from '@carbon/react';
+import {ProcessInstance} from '@vzeta/camunda-api-zod-schemas';
+import {Paths} from 'modules/Routes';
+import {useCancelProcessInstance} from 'modules/mutations/processInstance/useCancelProcessInstance';
+import {OperationItem} from 'modules/components/OperationItem';
+import {Restricted} from 'modules/components/Restricted';
+import {Link} from 'modules/components/Link';
+import {useRootInstanceId} from 'modules/queries/callHierarchy/useRootInstanceId';
+import {notificationsStore} from 'modules/stores/notifications';
+
+type Props = {
+  processInstanceKey: ProcessInstance['processInstanceKey'];
+  permissions?: ResourceBasedPermissionDto[] | null;
+};
+
+const Cancel: React.FC<Props> = ({processInstanceKey, permissions}) => {
+  const [isCancellationModalVisible, setIsCancellationModalVisible] =
+    useState(false);
+
+  const {data: rootInstanceId} = useRootInstanceId({
+    enabled: isCancellationModalVisible,
+  });
+
+  const cancelProcessInstanceMutation = useCancelProcessInstance(
+    processInstanceKey,
+    (error) =>
+      notificationsStore.displayNotification({
+        kind: 'error',
+        title: 'Failed to cancel process instance',
+        subtitle: error.message,
+        isDismissable: true,
+      }),
+  );
+
+  return (
+    <>
+      <Restricted
+        resourceBasedRestrictions={{
+          scopes: ['UPDATE_PROCESS_INSTANCE'],
+          permissions,
+        }}
+      >
+        <OperationItem
+          type="CANCEL_PROCESS_INSTANCE"
+          onClick={() => setIsCancellationModalVisible(true)}
+          title={`Cancel Instance ${processInstanceKey}`}
+          disabled={cancelProcessInstanceMutation.isPending}
+          size="sm"
+        />
+      </Restricted>
+
+      {isCancellationModalVisible && (
+        <>
+          {!rootInstanceId ? (
+            <Modal
+              open={isCancellationModalVisible}
+              preventCloseOnClickOutside
+              modalHeading="Apply Operation"
+              primaryButtonText="Apply"
+              secondaryButtonText="Cancel"
+              onRequestSubmit={() => {
+                setIsCancellationModalVisible(false);
+                cancelProcessInstanceMutation.mutate();
+              }}
+              onRequestClose={() => setIsCancellationModalVisible(false)}
+              size="md"
+              data-testid="confirm-cancellation-modal"
+            >
+              <p>{`About to cancel Instance ${processInstanceKey}. In case there are called instances, these will be canceled too.`}</p>
+              <p>Click "Apply" to proceed.</p>
+            </Modal>
+          ) : (
+            <Modal
+              open={isCancellationModalVisible}
+              preventCloseOnClickOutside
+              modalHeading="Cancel Instance"
+              passiveModal
+              onRequestClose={() => setIsCancellationModalVisible(false)}
+              size="md"
+              data-testid="passive-cancellation-modal"
+            >
+              <p>
+                To cancel this instance, the root instance{' '}
+                <Link
+                  to={Paths.processInstance(rootInstanceId)}
+                  title={`View root instance ${rootInstanceId}`}
+                >
+                  {rootInstanceId}
+                </Link>{' '}
+                needs to be canceled. When the root instance is canceled all the
+                called instances will be canceled automatically.
+              </p>
+            </Modal>
+          )}
+        </>
+      )}
+    </>
+  );
+};
+
+export {Cancel};
