+++ operate/client/src/modules/components/Operations/v2/index.tsx
@@ -8,25 +8,22 @@
 
 import React, {useState} from 'react';
 
-import {ACTIVE_OPERATION_STATES} from 'modules/constants';
 import {ErrorHandler, operationsStore} from 'modules/stores/operations';
 import {observer} from 'mobx-react';
 
 import {OperationItems} from 'modules/components/OperationItems';
 import {OperationItem} from 'modules/components/OperationItem';
-import {DangerButton} from 'modules/components/OperationItem/DangerButton';
 import {modificationsStore} from 'modules/stores/modifications';
 import {Restricted} from 'modules/components/Restricted';
-import {Modal, InlineLoading} from '@carbon/react';
-import {Paths} from 'modules/Routes';
-import {Link} from 'modules/components/Link';
+import {InlineLoading} from '@carbon/react';
 import {OperationsContainer} from '../styled';
 import {processInstancesStore} from 'modules/stores/processInstances';
 import {getStateLocally} from 'modules/utils/localStorage';
 import {ModificationHelperModal} from '../ModificationHelperModal';
 import {ProcessInstance} from '@vzeta/camunda-api-zod-schemas';
-import {useOperations} from 'modules/queries/operations/useOperations';
-import {useRootInstanceId} from 'modules/queries/callHierarchy/useRootInstanceId';
+import {Cancel} from './Cancel';
+import {Delete} from './Delete';
+import {ResolveIncident} from './ResolveIncident';
 
 type Props = {
   instance: ProcessInstance;
@@ -48,16 +45,10 @@ const Operations: React.FC<Props> = observer(
     isInstanceModificationVisible = false,
     permissions,
   }) => {
-    const [isDeleteModalVisible, setIsDeleteModalVisible] = useState(false);
-    const [isCancellationModalVisible, setIsCancellationModalVisible] =
-      useState(false);
     const [
       isModificationModeHelperModalVisible,
       setIsModificationModeHelperModalVisible,
     ] = useState(false);
-
-    const {data: operations} = useOperations();
-    const {data: rootInstanceId} = useRootInstanceId();
 
     const {isModificationModeEnabled} = modificationsStore;
 
@@ -76,15 +67,7 @@ const Operations: React.FC<Props> = observer(
       onOperation?.(operationType);
     };
 
-    const isOperationActive = (operationType: OperationEntityType) => {
-      return operations?.some(
-        (operation) =>
-          operation.type === operationType &&
-          ACTIVE_OPERATION_STATES.includes(operation.state),
-      );
-    };
-
-    const isRunning = instance.state === 'ACTIVE';
+    const isInstanceActive = instance.state === 'ACTIVE';
 
     return (
       <OperationsContainer orientation="horizontal">
@@ -99,56 +82,28 @@ const Operations: React.FC<Props> = observer(
         )}
         <OperationItems>
           {instance.hasIncident && !isModificationModeEnabled && (
-            <Restricted
-              resourceBasedRestrictions={{
-                scopes: ['UPDATE_PROCESS_INSTANCE'],
-                permissions,
-              }}
-            >
-              <OperationItem
-                type="RESOLVE_INCIDENT"
-                onClick={() => applyOperation('RESOLVE_INCIDENT')}
-                title={`Retry Instance ${instance.processInstanceKey}`}
-                disabled={isOperationActive('RESOLVE_INCIDENT')}
-                size="sm"
-              />
-            </Restricted>
+            <ResolveIncident
+              processInstanceKey={instance.processInstanceKey}
+              permissions={permissions}
+              applyOperation={applyOperation}
+            />
           )}
-          {isRunning && !isModificationModeEnabled && (
-            <Restricted
-              resourceBasedRestrictions={{
-                scopes: ['UPDATE_PROCESS_INSTANCE'],
-                permissions,
-              }}
-            >
-              <OperationItem
-                type="CANCEL_PROCESS_INSTANCE"
-                onClick={() => setIsCancellationModalVisible(true)}
-                title={`Cancel Instance ${instance.processInstanceKey}`}
-                disabled={isOperationActive('CANCEL_PROCESS_INSTANCE')}
-                size="sm"
-              />
-            </Restricted>
+          {isInstanceActive && !isModificationModeEnabled && (
+            <Cancel
+              processInstanceKey={instance.processInstanceKey}
+              permissions={permissions}
+            />
           )}
-          {!isRunning && (
-            <Restricted
-              resourceBasedRestrictions={{
-                scopes: ['DELETE_PROCESS_INSTANCE'],
-                permissions,
-              }}
-            >
-              <DangerButton
-                type="DELETE"
-                onClick={() => setIsDeleteModalVisible(true)}
-                title={`Delete Instance ${instance.processInstanceKey}`}
-                disabled={isOperationActive('DELETE_PROCESS_INSTANCE')}
-                size="sm"
-              />
-            </Restricted>
+          {!isInstanceActive && (
+            <Delete
+              processInstanceKey={instance.processInstanceKey}
+              permissions={permissions}
+              applyOperation={applyOperation}
+            />
           )}
 
           {isInstanceModificationVisible &&
-            isRunning &&
+            isInstanceActive &&
             !isModificationModeEnabled && (
               <Restricted
                 resourceBasedRestrictions={{
@@ -171,72 +126,6 @@ const Operations: React.FC<Props> = observer(
               </Restricted>
             )}
         </OperationItems>
-
-        {isCancellationModalVisible && (
-          <>
-            {!rootInstanceId ? (
-              <Modal
-                open={isCancellationModalVisible}
-                preventCloseOnClickOutside
-                modalHeading="Apply Operation"
-                primaryButtonText="Apply"
-                secondaryButtonText="Cancel"
-                onRequestSubmit={() => {
-                  setIsCancellationModalVisible(false);
-                  applyOperation('CANCEL_PROCESS_INSTANCE');
-                }}
-                onRequestClose={() => setIsCancellationModalVisible(false)}
-                size="md"
-                data-testid="confirm-cancellation-modal"
-              >
-                <p>{`About to cancel Instance ${instance.processInstanceKey}. In case there are called instances, these will be canceled too.`}</p>
-                <p>Click "Apply" to proceed.</p>
-              </Modal>
-            ) : (
-              <Modal
-                open={isCancellationModalVisible}
-                preventCloseOnClickOutside
-                modalHeading="Cancel Instance"
-                passiveModal
-                onRequestClose={() => setIsCancellationModalVisible(false)}
-                size="md"
-                data-testid="passive-cancellation-modal"
-              >
-                <p>
-                  To cancel this instance, the root instance{' '}
-                  <Link
-                    to={Paths.processInstance(rootInstanceId)}
-                    title={`View root instance ${rootInstanceId}`}
-                  >
-                    {rootInstanceId}
-                  </Link>{' '}
-                  needs to be canceled. When the root instance is canceled all
-                  the called instances will be canceled automatically.
-                </p>
-              </Modal>
-            )}
-          </>
-        )}
-        {isDeleteModalVisible && (
-          <Modal
-            open={isDeleteModalVisible}
-            danger
-            preventCloseOnClickOutside
-            modalHeading="Delete Instance"
-            primaryButtonText="Delete"
-            secondaryButtonText="Cancel"
-            onRequestSubmit={() => {
-              applyOperation('DELETE_PROCESS_INSTANCE');
-              setIsDeleteModalVisible(false);
-            }}
-            onRequestClose={() => setIsDeleteModalVisible(false)}
-            size="md"
-            data-testid="confirm-deletion-modal"
-          >
-            <p>About to delete Instance {instance.processInstanceKey}.</p>
-            <p>Click "Delete" to proceed.</p>
-          </Modal>
-        )}
 
         {isModificationModeHelperModalVisible && (
           <ModificationHelperModal
